requirements.txt
requests==2.31.0
pandas==2.2.2
google-cloud-bigquery==3.20.1
google-cloud-bigquery-storage==2.25.0
db-dtypes==1.2.0
python-dotenv==1.0.1
pytest==8.2.0
